+++ bindings/vue2/vue-onsenui/CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 ====
 
+dev
+---
+
+ ### BREAKING CHANGES
+
+ * ESM build is now distributed as untranspiled ES6. If you need ES5 code, use the UMD build (dist/vue-onsenui.js) or transpile the `esm` directory using a tool such as Babel.
+
+ ### Misc
+
+ * ESM build is now distributed as untranspiled ES6.
+
 2.6.4
 ---